+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 > and therefore its `version` is NOT changed when any updates are made to this template.
 > Instead, updates to this template are shown only in this CHANGELOG (and in the Git history).
 
+### 2022-10-21
+
+- Begin to use SQLite -- use the `fix_electron20_build` branch of a fork of `better-sqlite3`
+  because the mainstream version doesn't yet support Electron 20:
+  - https://github.com/WiseLibs/better-sqlite3/issues/867#issuecomment-1277766794
+  - https://github.com/WiseLibs/better-sqlite3/pull/870
+  - https://github.com/neoxpert/better-sqlite3/tree/fix_electron20_build
+
 ### 2022-10-16
 
 - Create the [`dotnet` branch](https://github.com/cwellsx/electron_forge_template/tree/dotnet)